refactor: move getSimpleObject into composables/utils.js
